requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
python-dotenv==1.1.0
playwright==1.52.0
apscheduler==3.10.4
pgvector==0.4.1
sentence-transformers==2.3.1
pydantic-settings==2.0.2
beautifulsoup4==4.13.4
robotexclusionrulesparser==1.7.1